requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.3.5